feat: 状态管理从Provider 全面迁移至Riverpod
