--- lib/ResponseCard/hero_dialog_route.dart
@@ -1,49 +1,0 @@
-import 'package:flutter/material.dart';
-
-/// {@template hero_dialog_route}
-/// Custom [PageRoute] that creates an overlay dialog (popup effect).
-///
-/// Best used with a [Hero] animation.
-/// {@endtemplate}
-/// Hero 卡片弹出动画的组件
-class HeroDialogRoute<T> extends PageRoute<T> {
-  /// {@macro hero_dialog_route}
-  HeroDialogRoute({
-    required WidgetBuilder builder,
-    // RouteSettings settings,
-    bool fullscreenDialog = false,
-  })  : _builder = builder,
-        super(fullscreenDialog: fullscreenDialog);
-
-  final WidgetBuilder _builder;
-
-  @override
-  bool get opaque => false;
-
-  @override
-  bool get barrierDismissible => true;
-
-  @override
-  Duration get transitionDuration => const Duration(milliseconds: 300);
-
-  @override
-  bool get maintainState => true;
-
-  @override
-  Color get barrierColor => Colors.black54;
-
-  @override
-  Widget buildTransitions(BuildContext context, Animation<double> animation,
-      Animation<double> secondaryAnimation, Widget child) {
-    return child;
-  }
-
-  @override
-  Widget buildPage(BuildContext context, Animation<double> animation,
-      Animation<double> secondaryAnimation) {
-    return _builder(context);
-  }
-
-  @override
-  String get barrierLabel => 'Popup dialog open';
-}
