+++ lib/main.dart
@@ -5,29 +5,19 @@ import 'package:flutter/material.dart';
 // utils
 import 'package:intl/date_symbol_data_local.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
-// models
-import 'package:namer_app/ResponseCard/response_card_model.dart';
 
 // pages
 import 'TodayList/today_list_page.dart';
 import 'pages/AnalysisReview/review_analysis_page.dart';
-import 'MyPage/my_page.dart';
+import 'pages/MyPage/my_page.dart';
 import 'DietaryAnalysis/dietary_analysis_page.dart';
 import 'pages/Login/login_page.dart';
 import 'pages/Login/register_info_page.dart';
-import 'package:provider/provider.dart' as provider;
 import 'debugButton.dart';
-import 'DailyDiet/event_log_page.dart';
+import 'pages/DailyDiet/event_log_page.dart';
 
 // riverpod 状态管理
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-
-// part 'main.g.dart';
-
-final responseCardProvider =
-    ChangeNotifierProvider<ResponseCardModelProvider>((ref) {
-  return ResponseCardModelProvider();
-});
 
 // 用于控制全局路由
 final GlobalKey<NavigatorState> navigatorKey = GlobalKey<NavigatorState>();
@@ -50,65 +40,57 @@ void main() {
 class MyApp extends ConsumerWidget {
   @override
   Widget build(BuildContext context, WidgetRef ref) {
-    return provider.MultiProvider(
-      // 使用MultiProvider包裹MaterialApp
-      providers: [
-        provider.ChangeNotifierProvider.value(
-            value: ref.watch(responseCardProvider.notifier)), // 创建冲动响应卡实例
+    return MaterialApp(
+      title: 'CBT-E App',
+      // 配置本地化
+      localizationsDelegates: [
+        GlobalMaterialLocalizations.delegate, // Material 组件本地化
+        GlobalCupertinoLocalizations.delegate, // Cupertino 组件本地化
+        GlobalWidgetsLocalizations.delegate, // 基本小部件本地化
+      ],
+      supportedLocales: [
+        Locale('zh', ''), // 中文
+        Locale('en', ''), // 英文
+        Locale('es', ''), // 西班牙语（可根据需要添加更多语言）
       ],
 
-      child: MaterialApp(
-          title: 'CBT-E App',
-          // 配置本地化
-          localizationsDelegates: [
-            GlobalMaterialLocalizations.delegate, // Material 组件本地化
-            GlobalCupertinoLocalizations.delegate, // Cupertino 组件本地化
-            GlobalWidgetsLocalizations.delegate, // 基本小部件本地化
-          ],
-          supportedLocales: [
-            Locale('zh', ''), // 中文
-            Locale('en', ''), // 英文
-            Locale('es', ''), // 西班牙语（可根据需要添加更多语言）
-          ],
-
-          // 设置默认语言（可根据用户设备的语言设置）
-          locale: Locale('zh', ''), // 默认中文
-          home: MainScreen(),
-          theme: ThemeData(
-            dialogTheme: DialogTheme(
-              backgroundColor: Colors.white, // 设置弹窗的全局背景色为白色
-              shape: RoundedRectangleBorder(
-                borderRadius: BorderRadius.circular(10), // 可选：为弹窗添加圆角
-              ),
-            ),
-            elevatedButtonTheme: ElevatedButtonThemeData(
-              style: ElevatedButton.styleFrom(
-                backgroundColor: Color(0xFFD98E8E), // 设置按钮背景色
-                foregroundColor: Colors.white, // 设置按钮文字颜色
-                shadowColor: Colors.grey, // 设置阴影颜色
-                elevation: 5, // 设置阴影高度
-                shape: RoundedRectangleBorder(
-                  borderRadius: BorderRadius.circular(24), // 设置按钮的圆角
-                ),
-              ),
-            ),
-            appBarTheme: AppBarTheme(backgroundColor: Color(0xFFF0E5E7)),
-            colorScheme: ColorScheme.fromSeed(
-              seedColor: Color(0xFFD98E8E), // 设置全局的主颜色
-              primary: Color(0xFFD98E8E), // 主要按钮颜色
-              onPrimary: Color(0xFFFFFFFF), // 主按钮下的字色
-              secondary: Color(0xFFFFFFFF), // 次要按钮颜色
-              onSecondary: Color(0xFF505050), // 普通按钮字色
-            ),
-            scaffoldBackgroundColor: Color(0xFFF0E5E7), // 设置整体背景色
+      // 设置默认语言（可根据用户设备的语言设置）
+      locale: Locale('zh', ''), // 默认中文
+      home: MainScreen(),
+      theme: ThemeData(
+        dialogTheme: DialogTheme(
+          backgroundColor: Colors.white, // 设置弹窗的全局背景色为白色
+          shape: RoundedRectangleBorder(
+            borderRadius: BorderRadius.circular(10), // 可选：为弹窗添加圆角
           ),
-          initialRoute: '/',
-          routes: {
-            '/login': (context) => LoginPage(), // 登录页面
-            '/home': (context) => MainScreen(), // 主屏幕，登录成功后跳转的页面
-            '/register': (context) => RegisterInfoPage(), // 添加注册页面的路由
-          },
-          navigatorKey: navigatorKey),
+        ),
+        elevatedButtonTheme: ElevatedButtonThemeData(
+          style: ElevatedButton.styleFrom(
+            backgroundColor: Color(0xFFD98E8E), // 设置按钮背景色
+            foregroundColor: Colors.white, // 设置按钮文字颜色
+            shadowColor: Colors.grey, // 设置阴影颜色
+            elevation: 5, // 设置阴影高度
+            shape: RoundedRectangleBorder(
+              borderRadius: BorderRadius.circular(24), // 设置按钮的圆角
+            ),
+          ),
+        ),
+        appBarTheme: AppBarTheme(backgroundColor: Color(0xFFF0E5E7)),
+        colorScheme: ColorScheme.fromSeed(
+          seedColor: Color(0xFFD98E8E), // 设置全局的主颜色
+          primary: Color(0xFFD98E8E), // 主要按钮颜色
+          onPrimary: Color(0xFFFFFFFF), // 主按钮下的字色
+          secondary: Color(0xFFFFFFFF), // 次要按钮颜色
+          onSecondary: Color(0xFF505050), // 普通按钮字色
+        ),
+        scaffoldBackgroundColor: Color(0xFFF0E5E7), // 设置整体背景色
+      ),
+      initialRoute: '/',
+      routes: {
+        '/login': (context) => LoginPage(), // 登录页面
+        '/home': (context) => MainScreen(), // 主屏幕，登录成功后跳转的页面
+        '/register': (context) => RegisterInfoPage(), // 添加注册页面的路由
+      },
     );
   }
 }
